backend/requirements.txt
fastapi
uvicorn
motor
pymongo
passlib[bcrypt]
python-jose[cryptography]
email-validator
python-multipart    # 👈 this line solves your backend error
